Rename cache module to hybridCache
