+++ src/utils/caching/hybridCache.test.ts
@@ -1,8 +1,8 @@
-import cache from '../memoization/cache';
-import { reset as inDiskCacheReset } from '../memoization/inFileCacheStorage';
-import { reset as inMemoryCacheReset } from '../memoization/inMemoryCacheStorage';
 import waitFor from '../timers/waitFor';
-import { StorageOptions } from './cacheStorage';
+import cache from './hybridCache';
+import { StorageOptions } from './hybridCacheStorage';
+import { reset as inDiskCacheReset } from './inDiskCacheStorage';
+import { reset as inMemoryCacheReset } from './inMemoryCacheStorage';
 
 describe(`utils/memoization/cache.ts`, () => {
   beforeEach(() => {
@@ -20,8 +20,8 @@ describe(`utils/memoization/cache.ts`, () => {
 
       describe(`when using in-memory storage`, () => {
         test(`when using the default TTL`, async () => {
-          const cacheHitsBefore = require('../memoization/inMemoryCacheStorage').cacheHits;
-          const cacheMissBefore = require('../memoization/inMemoryCacheStorage').cacheMiss;
+          const cacheHitsBefore = require('.//inMemoryCacheStorage').cacheHits;
+          const cacheMissBefore = require('.//inMemoryCacheStorage').cacheMiss;
           expect(await cache('key', async () => {
             await waitFor(1);
             return expectationResult;
@@ -31,8 +31,8 @@ describe(`utils/memoization/cache.ts`, () => {
             return Promise.resolve(expectationResult);
           })).toEqual(expectationResult);
 
-          const cacheHitsAfter = require('../memoization/inMemoryCacheStorage').cacheHits;
-          const cacheMissAfter = require('../memoization/inMemoryCacheStorage').cacheMiss;
+          const cacheHitsAfter = require('.//inMemoryCacheStorage').cacheHits;
+          const cacheMissAfter = require('.//inMemoryCacheStorage').cacheMiss;
           expect(cacheHitsAfter).toBeGreaterThan(cacheHitsBefore);
           expect(cacheMissAfter).toEqual(cacheMissBefore + 1); // Cache should have been missed only for the first call
         });
@@ -41,8 +41,8 @@ describe(`utils/memoization/cache.ts`, () => {
           const expectationResult = { key2: 'value2' };
 
           test(`when using TTL of 1 second and waiting more than 1 second between calls`, async () => {
-            const cacheHitsBefore = require('../memoization/inMemoryCacheStorage').cacheHits;
-            const cacheMissBefore = require('../memoization/inMemoryCacheStorage').cacheMiss;
+            const cacheHitsBefore = require('.//inMemoryCacheStorage').cacheHits;
+            const cacheMissBefore = require('.//inMemoryCacheStorage').cacheMiss;
             await waitFor(1001);
             expect(await cache('key2', async () => {
               await waitFor(1);
@@ -54,8 +54,8 @@ describe(`utils/memoization/cache.ts`, () => {
               return Promise.resolve(expectationResult);
             })).toEqual(expectationResult);
 
-            const cacheHitsAfter = require('../memoization/inMemoryCacheStorage').cacheHits;
-            const cacheMissAfter = require('../memoization/inMemoryCacheStorage').cacheMiss;
+            const cacheHitsAfter = require('.//inMemoryCacheStorage').cacheHits;
+            const cacheMissAfter = require('.//inMemoryCacheStorage').cacheMiss;
             expect(cacheHitsAfter).toEqual(cacheHitsBefore + 1);
             expect(cacheMissAfter).toBeGreaterThan(cacheMissBefore);
           });
