+++ src/utils/caching/inDiskCacheStorage.ts
@@ -1,7 +1,7 @@
+import * as Sentry from '@sentry/node';
 import path from 'path';
 import { deleteFile, readFile, writeFile } from '../node/fs-utils';
-import { CachedItem, CacheStorage as GenericCacheStorage, Get, Reset, Set, StorageOptions as GenericStorageOptions } from './cacheStorage';
-import * as Sentry from '@sentry/node';
+import { CachedItem, Get, HybridCacheStorage as GenericCacheStorage, Reset, Set, StorageOptions as GenericStorageOptions } from './hybridCacheStorage';
 
 type StorageOptions = GenericStorageOptions<{ filename: string }>;
 type CacheStorage = GenericCacheStorage<any, StorageOptions>;
